LF1 can be stepped with AOM while locked to TCL
Now that the 899 is locked to tcl, an AOM is used to derive F=0, and
stepped in hardware to produce the LF1 channel. FLPZT is renamed to 899
external scan, but no longer controlled through Harware control (Though
TCL and Scan Master can access it to lock/scan the laser). FLPZT2 is the
Menlo laser and is renamed MenloPZT. Generally tidy things up because the
code involving the lasers had become very confusing. Remove the pump AOM
steps, eventually the Pump will also be passed through an AOM and stepped
intdependently. Retain a lot of now redundant code in HC so that it won't
be too hard to return to the I2 lock or run a nufern if we need to